minor fixes and build configs

+++ WillEngine/src/engine/vulkan/vk_loader.cpp
@@ -280,7 +280,7 @@ std::optional<std::shared_ptr<LoadedGLTF>> loadGltf(VulkanEngine* engine, std::s
 
 				fastgltf::iterateAccessor<std::uint32_t>(gltf, indexaccessor,
 					[&](std::uint32_t idx) {
-						indices.push_back(idx + initial_vtx);
+						indices.push_back(idx + static_cast<uint32_t>(initial_vtx));
 					});
 			}
